[end ch1365] Moving date picker to top. Adding Quick Dates icon & dropdown.
