+++ src/components/Dashboard/Dashboard.js
@@ -1,9 +1,13 @@
-import React from "react"
+import React, { useState, forwardRef } from "react"
 import "App.css"
 import NavHeader from "components/Nav/NavHeader"
 import PageTitle from "components/Nav/PageTitle"
 import KeyMetricsGraph from "components/Dashboard/KeyMetricsGraph"
+import StaredDates from "components/Dashboard/StaredDates"
 import { subDays } from "date-fns"
+import DatePicker from "react-datepicker"
+import "react-datepicker/dist/react-datepicker.css"
+import { ReactSVG } from "react-svg"
 
 export default function Dashboard() {
   // data fixture
@@ -25,11 +29,44 @@ export default function Dashboard() {
     })
   }
 
+  // date picker
+  const today = new Date()
+  const prevDay = new Date()
+  const [dateRange, setDateRange] = useState([
+    prevDay.setDate(today.getDate() - 30),
+    new Date(),
+  ])
+
+  const [startDate, endDate] = dateRange
+  const ExampleCustomInput = forwardRef(({ value, onClick }, ref) => (
+    <button className='' onClick={onClick} ref={ref}>
+      {value}
+      <ReactSVG
+        src='/graphics/utiCarretDown.svg'
+        className='svg-primary inline-block align-middle ml-1'
+      />
+    </button>
+  ))
+
   return (
     <div className='App'>
       <NavHeader Position='top' />
-      <div className='mt-20 pl-4'>
+      <div className='mt-20 px-4 flex justify-between flex-nowrap'>
         <PageTitle title='Dashboard' />
+
+        <div className='flex flex-nowrap'>
+          <DatePicker
+            className='cursor-pointer'
+            selectsRange={true}
+            startDate={startDate}
+            endDate={endDate}
+            onChange={(update) => {
+              setDateRange(update)
+            }}
+            customInput={<ExampleCustomInput />}
+          />
+          <StaredDates />
+        </div>
       </div>
       <main className='p-4'>
         <KeyMetricsGraph data={data} />
